+++ jscomp/ext/literals.ml
@@ -130,10 +130,10 @@ let esmodule = "esmodule"
 let commonjs = "commonjs"
 
 let es6 = "es6"
-[@@ocaml.deprecated "Will be removed in v12"]
+(* [@@deprecated "Will be removed in v12"] *)
 
 let es6_global = "es6-global"
-[@@ocaml.deprecated "Will be removed in v12"]
+(* [@@deprecated "Will be removed in v12"] *)
 
 let unused_attribute = "Unused attribute "